requirements.txt
llama-index==0.9.48
openai==1.3.6
python-dotenv==1.0.0
markdown==3.5.1
pyyaml==6.0.1